Feature/star and long in entities

+++ src/Company.SharedKernel/Handlers/UpdateAggregateRootCommand.cs
@@ -1,4 +1,4 @@
 ﻿namespace Company.SharedKernel.Abstractions;
 
-public record UpdateAggregateRootCommand<TAggregateRoot>(long Id) : IRequest<Result<TAggregateRoot>>
+public record UpdateAggregateRootCommand<TAggregateRoot>(int Id) : IRequest<Result<TAggregateRoot>>
     where TAggregateRoot : class, IAggregateRoot;
